improve training
if the standard deviation decays low enough, then the model freaks out and forgets everything. this is because the learning rates are too high for lower standard deviations, and have to be decayed in addition to the standard deviation

this implementation of PPO decays standard deviation, but other implementations do something different. i think i may want to research what they're doing to see if it leads to more effective/consistent training